config refactored,  refactored
