+++ app/config/extension.doctrine.neon
@@ -1,0 +1,13 @@
+extensions:
+  events: Kdyby\Events\DI\EventsExtension
+  annotations: Kdyby\Annotations\DI\AnnotationsExtension
+  doctrine: Kdyby\Doctrine\DI\OrmExtension
+  doctrine2identity: Majkl578\NetteAddons\Doctrine2Identity\DI\IdentityExtension
+
+doctrine:
+  host: %database.host%
+  user: %database.user%
+  password: %database.password%
+  dbname: %database.dbname%
+  metadata:
+    Supervizor: %appDir%
