+++ app/config/extension.importer.neon
@@ -1,0 +1,17 @@
+extensions:
+  importer: Supervizor\Utils\Importer\ImporterExtension
+
+importer:
+  target: Supervizor\Import\ImportTarget
+  imports:
+    mf:
+      title: "Ministerstvo financí"
+      default: true
+      datasets:
+        mf2016:
+          title: "Rok 2016"
+          description: "Přehled faktur Ministerstva financí"
+          source: "http://data.mfcr.cz/cs/api/3/action/resource_show?id=aec18a6a-0d8f-49a4-a8e7-ae0fbd32125f"
+          homepage: "http://data.mfcr.cz/cs/dataset/prehled-faktur-ministerstva-financi-cr"
+          parser: Supervizor\Utils\Importer\Parsers\Mfcr
+          default: true
