+++ app/config/extension.nette.neon
@@ -1,0 +1,15 @@
+php:
+  date.timezone: Europe/Prague
+
+nette:
+  application:
+    errorPresenter: Error
+    mapping:
+      *: Supervizor\Application\*\*Presenter
+
+  session:
+    expiration: 14 days
+
+services:
+  routerFactory: Supervizor\UI\RouterFactory
+  router: @routerFactory::createRouter
