--- app/config/services.neon
@@ -1,33 +1,0 @@
-
-services:
-    - Nextras\Forms\Rendering\Bs3FormRenderer
-    - Supervizor\Settings\SupervizorSettings
-
-    # Compoenents
-    - Supervizor\UI\BaseForm
-    - Supervizor\Auth\Controls\SignInFormFactory
-
-    # security
-    - Supervizor\Auth\Authenticator
-    - Supervizor\Auth\PasswordManager
-
-    # repository
-    - Supervizor\Auth\UserRepository
-    - Supervizor\Supplies\SupplierRepository
-    - Supervizor\Invoice\InvoiceRepository
-    - Supervizor\Budget\BudgetRepository
-    - Supervizor\Import\ImportRepository
-
-    # import
-    - 
-        class: Supervizor\Import\ImportTarget
-        inject: true
-
-    # commands
-    -
-        class: Supervizor\Import\Defaults\DefaultData
-        tags: [kdyby.console.command]
-    
-    -
-        class: Supervizor\Console\Commands\Import
-        tags: [kdyby.console.command]
