+++ composer.json
@@ -10,6 +10,7 @@
   "require": {
     "php": ">= 5.3.0",
     "janmarek/webloader": "dev-master",
+    "joseki/webloader-filters": "~1.0",
     "leafo/scssphp": "dev-master",
     "natxet/CssMin": "dev-master",
     "dg/adminer-custom": "dev-master",
